freetype: Prevent lose of data when fixed point divide calculation.
Summary:
In environments that long is 4byte, fixed-point division calculations will cause data loss.
fixed-point division need to more space.
Therefore, change all long types to long long types.

Test Plan: N/A

Reviewers: Hermet, kimcinoo, smohanty

Subscribers: vtorri, cedric, #reviewers, #committers

Tags: #efl

Differential Revision: https://phab.enlightenment.org/D9817

+++ src/static_libs/freetype/sw_ft_types.h
@@ -10,7 +10,7 @@
 /*    This type is used to store 16.16 fixed-point values, like scaling  */
 /*    values or matrix coefficients.                                     */
 /*                                                                       */
-typedef signed long  SW_FT_Fixed;
+typedef signed long long SW_FT_Fixed;
 
 
 /*************************************************************************/
